Move audio queueing from espeak/espeak-ng to shared file
So it can be used by other modules too.
